move : 백 <-> UI 타입 매핑 파일 분리 및 기존 중복 코드 삭제
